Split symbol table behavior edges
